+++ cmd/admin-config-get.go
@@ -20,7 +20,7 @@ import (
 	"encoding/json"
 
 	"github.com/minio/cli"
-	"github.com/minio/minio/pkg/probe"
+	"github.com/minio/mc/pkg/probe"
 )
 
 var adminConfigGetCmd = cli.Command{
