+++ cmd/admin-config-set.go
@@ -24,8 +24,8 @@ import (
 	"github.com/fatih/color"
 	"github.com/minio/cli"
 	"github.com/minio/mc/pkg/console"
+	"github.com/minio/mc/pkg/probe"
 	"github.com/minio/minio/pkg/madmin"
-	"github.com/minio/minio/pkg/probe"
 )
 
 var adminConfigSetCmd = cli.Command{
