+++ cmd/admin-credentials.go
@@ -18,7 +18,7 @@ package cmd
 
 import (
 	"github.com/minio/cli"
-	"github.com/minio/minio/pkg/probe"
+	"github.com/minio/mc/pkg/probe"
 )
 
 var (
