+++ cmd/admin-lock-clear.go
@@ -23,8 +23,8 @@ import (
 	"time"
 
 	"github.com/minio/cli"
+	"github.com/minio/mc/pkg/probe"
 	"github.com/minio/minio/pkg/madmin"
-	"github.com/minio/minio/pkg/probe"
 )
 
 var (
