+++ cmd/admin-service-restart.go
@@ -22,7 +22,7 @@ import (
 	"github.com/fatih/color"
 	"github.com/minio/cli"
 	"github.com/minio/mc/pkg/console"
-	"github.com/minio/minio/pkg/probe"
+	"github.com/minio/mc/pkg/probe"
 )
 
 var adminServiceRestartCmd = cli.Command{
