+++ cmd/cat-main.go
@@ -25,7 +25,7 @@ import (
 	"syscall"
 
 	"github.com/minio/cli"
-	"github.com/minio/minio/pkg/probe"
+	"github.com/minio/mc/pkg/probe"
 )
 
 var (
