+++ cmd/config.go
@@ -23,7 +23,7 @@ import (
 	"runtime"
 
 	"github.com/minio/go-homedir"
-	"github.com/minio/minio/pkg/probe"
+	"github.com/minio/mc/pkg/probe"
 )
 
 // mcCustomConfigDir contains the whole path to config dir. Only access via get/set functions.
