+++ cmd/pipe-main.go
@@ -21,7 +21,7 @@ import (
 	"syscall"
 
 	"github.com/minio/cli"
-	"github.com/minio/minio/pkg/probe"
+	"github.com/minio/mc/pkg/probe"
 )
 
 var (
