+++ cmd/session-migrate.go
@@ -21,7 +21,7 @@ import (
 	"strconv"
 
 	"github.com/minio/mc/pkg/console"
-	"github.com/minio/minio/pkg/probe"
+	"github.com/minio/mc/pkg/probe"
 	"github.com/minio/minio/pkg/quick"
 )
