+++ cmd/utils.go
@@ -27,7 +27,7 @@ import (
 	"time"
 
 	"github.com/minio/mc/pkg/console"
-	"github.com/minio/minio/pkg/probe"
+	"github.com/minio/mc/pkg/probe"
 )
 
 func isErrIgnored(err *probe.Error) (ignored bool) {
